Add Economic Paper of Nervos CKB
What kinds of token mechanisms are needed for securing a public blockchain? 
How can the token capture the value of the assets stored in this public blockchain platform?
What's gonna happen if a crypto-kitty worth $100,000,000 while the public blockchain that stored this kitty only worth $100,000?

This paper introduces a framework for analyzing these problems.
Also, a mechanism called Nervos CKB is presented for solving these problems.